FAT2-410 - Change so that all providers are returned
No longer filter out main providers and active providers when returning a provider or a list of providers
