+++ src/SFA.DAS.CourseDelivery.Application/Provider/Services/ProviderService.cs
@@ -28,7 +28,7 @@ namespace SFA.DAS.CourseDelivery.Application.Provider.Services
 
         public async Task<IEnumerable<ProviderSummary>> GetRegisteredProviders()
         {
-            var providersFromRepo = await _providerRegistrationRepository.GetAllRegisteredApprovedAndActiveProviders();
+            var providersFromRepo = await _providerRegistrationRepository.GetAllProviders();
 
             return providersFromRepo.Select(provider => (ProviderSummary)provider);
         }
@@ -119,14 +119,7 @@ namespace SFA.DAS.CourseDelivery.Application.Provider.Services
 
         public async Task<ProviderSummary> GetProviderByUkprn(int ukprn)
         {
-            var provider = await _providerRepository.GetByUkprn(ukprn);
-
-            if (provider == null)
-            {
-                return await  _providerRegistrationRepository.GetRegisteredApprovedAndActiveProviderByUkprn(ukprn);
-            }
-
-            return provider;
+            return await  _providerRegistrationRepository.GetProviderByUkprn(ukprn);
         }
 
         public async Task<UkprnsForStandard> GetUkprnsForStandardAndLocation(int standardId, double lat, double lon)
