+++ src/SFA.DAS.CourseDelivery.Domain/Interfaces/IProviderRegistrationRepository.cs
@@ -9,7 +9,7 @@ namespace SFA.DAS.CourseDelivery.Domain.Interfaces
         void DeleteAll();
         Task InsertMany(IEnumerable<ProviderRegistration> providerRegistrations);
         Task UpdateAddressesFromImportTable();
-        Task<ProviderRegistration> GetRegisteredApprovedAndActiveProviderByUkprn(int ukprn);
-        Task<IEnumerable<ProviderRegistration>> GetAllRegisteredApprovedAndActiveProviders();
+        Task<ProviderRegistration> GetProviderByUkprn(int ukprn);
+        Task<IEnumerable<ProviderRegistration>> GetAllProviders();
     }
 }
